Update Jelix 1.6.8-pre

+++ lib/jelix/plugins/cache/redis/redis.cache.php
@@ -51,6 +51,20 @@ class redisCacheDriver implements jICacheDriver {
     protected $key_prefix = '';
 
     /**
+     * method to flush the keys when key_prefix is used
+     *
+     * direct: uses SCAN and DEL, but it can take huge time
+     * jcacheredisworker: it stores the keys prefix to delete into a redis list
+     *     named 'jcacheredisdelkeys'.
+     *     You can use a script to launch a worker which pops from this list
+     *     prefix of keys to delete, and delete them with SCAN/DEL redis commands.
+     *     See the redisworker controller in the jelix module.
+     * event: send a jEvent. It's up to your application to respond to this event
+     *     and to implement your prefered method to delete all keys.
+     */
+    protected $key_prefix_flush_method = 'direct';
+
+    /**
      * @param Redis the redis connection
      */
     protected $redis;
@@ -87,12 +101,27 @@ class redisCacheDriver implements jICacheDriver {
             $this->key_prefix = $params['key_prefix'];
         }
 
+        if ($this->key_prefix && isset($params['key_prefix_flush_method'])) {
+            if (in_array($params['key_prefix_flush_method'],
+                         array('direct', 'jcacheredisworker', 'event'))) {
+                $this->key_prefix_flush_method = $params['key_prefix_flush_method'];
+            }
+        }
+
         // OK, let's connect now
         $this->redis = new Redis($params['host'], $params['port']);
 
         if (isset($params['db']) && intval($params['db']) != 0) {
             $this->redis->select_db($params['db']);
         }
+    }
+
+    /**
+     * Returns the redis api
+     * @return Redis
+     */
+    public function getRedis() {
+        return $this->redis;
     }
 
     /**
@@ -220,7 +249,7 @@ class redisCacheDriver implements jICacheDriver {
 
     /**
     * remove from the cache data of which TTL was expired
-    * element with TTL expired already removed => Nothing to do because memcache have an internal garbage mechanism
+    * element with TTL expired already removed => Nothing to do because redis has an internal garbage mechanism
     * @return boolean
     */
     public function garbage() {
@@ -228,11 +257,30 @@ class redisCacheDriver implements jICacheDriver {
     }
 
     /**
-    * clear all data in the cache
+    * clear all data in the cache.
+    *
+    * If key_prefix is set, only keys with that prefix will be removed.
+    * Note that in that case, it can result in a huge performance issue.
+    * See key_prefix_flush_method to configure the best method for your
+    * app and your server.
     * @return boolean       false if failure
     */
     public function flush() {
-        return ($this->redis->flushall()  == 'OK');
+        if (!$this->key_prefix) {
+            return ($this->redis->flushall()  == 'OK');
+        }
+        switch($this->key_prefix_flush_method) {
+            case 'direct':
+                $this->redis->flushByPrefix($this->key_prefix);
+                return true;
+            case 'event':
+                jEvent::notify('jCacheRedisFlushKeyPrefix', array('prefix'=>$this->key_prefix,
+                                                                  'profile' =>$this->profileName));
+                return true;
+            case 'jcacheredisworker':
+                $this->redis->rpush('jcacheredisdelkeys', $this->key_prefix);
+                return true;
+        }
     }
 
     protected function getUsedKey($key) {
